requirements.txt
torch==1.13.1
torchvision==0.14.1
open_clip_torch==2.20.0
Pillow==9.4.0
numpy==1.23.5
